fix: Fix AlbumByName to paginate over Google results
